dec2bin.c

+++ dec2bin.c
@@ -5,16 +5,16 @@
  */
 #include <stdio.h>
 
-void bin(unsigned long int num);
+void bin(long long int num);
 int main(void){
-    int num;
+    long long int num;
     printf("Digite um decimal de no maximo 64 bits: ");
     scanf("%llu", &num);
     bin(num);
  
     return 0;
 }
-void bin(unsigned long int num){
+void bin(long long int num){
     if(num != 0){    
         bin(num/2);
         printf("%d", num % 2);
